# Reduce thread locking behaviour to dial phase only