:wave: Back again with another controversial change. :smile: This aims to eliminate the bottleneck described in https://github.com/mdlayher/netlink/issues/154. This is the approach I described in https://github.com/mdlayher/netlink/issues/154#issuecomment-578045314. With the goal of dropping support for Go 1.11 and lower, this might be a good time to revisit.

To Do:
- [ ] Reintroduce a test for machines with network namespaces disabled, since `newLockedNetNSGoroutine` was removed

---

There are 2 elements at play here:
- The message misrouting discussed in http://lists.infradead.org/pipermail/libnl/2017-February/002293.html
  With this change, I'd like to start spelunking older versions of the kernel and the Go runtime for this bug. @mdlayher are you positive the `TestIntegrationConnConcurrent*` family of tests is able to reproduce this? Can you remember which Go and kernel version you were using at the time when this was written?

- Socket creation inside network namespaces
  As far as namespaces are concerned, OS threads only briefly need to enter a netns to create the netlink socket and can safely be returned to the host namespace once the socket fd has been allocated. We don't need to lock OS threads permanently to operate in multiple namespaces.

---

Some feedback please. :pray:  Still looking for a way to inject a mock `func getNS` somewhere into the `dial()` code path to emulate systems with network namespaces disabled.