@ti-mo 2020-09-11
Sorry, I haven't been around to work on this. Initial results are promising! The integration test suite completes about twice as fast on my machine, reflected in the profiles by an absence of lock contention around the worker closure channel. (since it's gone, obviously)

I'll do some virtme spelunking against older kernels and see if I can reproduce this thing. Take care!