@ti-mo 2020-09-12
Alright, I went as far back as kernel `3.2.0-5-amd64` that shipped with Debian Wheezy, integration test suite ran without any hiccups. Test binaries built with `go1.15.2`. All is well!