@ti-mo 2020-09-12
Timing the execution of the test suite on an 8-core Ryzen 3700X running kernel `5.7.19-1-ck`. Test binary built with go1.15.2.

Before: `sudo ./netlink.test -test.v  29.23s user 21.35s system 1299% cpu 3.892 total`
After: `sudo ./netlink.test -test.v  4.29s user 0.79s system 532% cpu 0.955 total`

These are not one-off results, the execution time sat consistently around the values shown.

Compatibility-wise, the integration test suite passes on systems as early as Linux `3.2.0-5-amd64`
with test binaries built with Go 1.10.