@ti-mo 2020-09-12
> To clarify, would you intend for this to go into a v1.2.0 (Go 1.12+ only) and _not_ v1.1.1 (Go 1.11+)?

According to the test suite, this works with Go 1.10 just fine, so not sure if we should worry too much about this.

Re: mocking a test for disabled namespaces: `fileNetNS()` doesn't have tests right now, which we should probably address. Still thinking about how to test `Dial()`. Any ideas?