@ti-mo 2020-09-17
@mdlayher I've added a short `netns_linux_test.go` to validate the behaviour of `netns.disabled`. Along with the existing integration test suite, that brings coverage of `netns_linux.go` to 100%. Feel free to merge if no more feedback from your side!